+++ src/java/beans/GenreList.java
@@ -25,7 +25,7 @@ public class GenreList {
             while (rs.next()) {
                 Genre genre = new Genre();
                 genre.setName(rs.getString("name"));
-                genre.setId((int) rs.getLong("id"));
+                genre.setId(rs.getLong("id"));
                 genreList.add(genre);
             }
